pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf2md"
version = "1.0.0"
description = "Enterprise-grade PDF-to-Markdown converter with multi-cloud storage, RBAC, and security features"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "PDF2MD Contributors"}
]
keywords = ["pdf", "markdown", "converter", "llm", "token-counting", "enterprise"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup",
    "Topic :: Utilities",
]

dependencies = [
    "pdfplumber>=0.11.0",
    "PyMuPDF>=1.23.0",
    "tiktoken>=0.5.0",
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.18",
    "bcrypt>=4.1.2",
    "aiosqlite>=0.19.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
# Cloud storage providers (install as needed)
azure = [
    "azure-storage-blob>=12.19.0",
    "azure-cosmos>=4.5.1",
]
gcp = [
    "google-cloud-storage>=2.14.0",
    "google-cloud-firestore>=2.14.0",
]
aws = [
    "boto3>=1.34.0",
]
all-cloud = [
    "azure-storage-blob>=12.19.0",
    "azure-cosmos>=4.5.1",
    "google-cloud-storage>=2.14.0",
    "google-cloud-firestore>=2.14.0",
    "boto3>=1.34.0",
]

dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "types-pyyaml>=6.0.0",
    "httpx>=0.26.0",
]

[project.scripts]
pdf2md = "pdf2md.cli.main:app"

[project.urls]
Homepage = "https://github.com/trumb/pdf-to-markdown"
Repository = "https://github.com/trumb/pdf-to-markdown"
Issues = "https://github.com/trumb/pdf-to-markdown/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pdf2md = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=pdf2md",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"tests/*" = ["S101"]  # assert allowed in tests

[tool.black]
target-version = ["py311"]
line-length = 100
include = '\.pyi?$'